+++ thoth/src/main/java/com/salesforce/apollo/thoth/grpc/dht/DhtClient.java
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-package com.salesforce.apollo.thoth.grpc;
+package com.salesforce.apollo.thoth.grpc.dht;
 
 import java.io.IOException;
 import java.util.List;
@@ -41,6 +41,7 @@ import com.salesforce.apollo.comm.ServerConnectionCache.ManagedServerConnection;
 import com.salesforce.apollo.crypto.Digest;
 import com.salesforce.apollo.membership.Member;
 import com.salesforce.apollo.stereotomy.services.grpc.StereotomyMetrics;
+import com.salesforce.apollo.stereotomy.services.proto.ProtoKERLService;
 
 /**
  * @author hal.hildebrand
@@ -54,7 +55,7 @@ public class DhtClient implements DhtService {
         };
     }
 
-    public static DhtService getLocalLoopback(Dht service, Member member) {
+    public static DhtService getLocalLoopback(ProtoKERLService service, Member member) {
         return new DhtService() {
 
             @Override
