+++ thoth/src/main/java/com/salesforce/apollo/thoth/grpc/validator/ValidatorServer.java
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-package com.salesforce.apollo.thoth.grpc;
+package com.salesforce.apollo.thoth.grpc.validator;
 
 import java.util.concurrent.Executor;
 
@@ -20,6 +20,7 @@ import com.salesforce.apollo.comm.RoutableService;
 import com.salesforce.apollo.crypto.Digest;
 import com.salesforce.apollo.stereotomy.services.grpc.StereotomyMetrics;
 import com.salesforce.apollo.thoth.Sakshi;
+import com.salesforce.apollo.thoth.grpc.reconciliation.ReconciliationServer;
 import com.salesforce.apollo.utils.Utils;
 
 import io.grpc.stub.StreamObserver;
