+++ thoth/src/main/java/com/salesforce/apollo/thoth/grpc/validator/ValidatorService.java
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-package com.salesforce.apollo.thoth.grpc;
+package com.salesforce.apollo.thoth.grpc.validator;
 
 import java.util.List;
